Move code that computes in-range expression to a new module
