+++ kani-compiler/src/codegen_cprover_gotoc/codegen/intrinsic.rs
@@ -7,7 +7,6 @@ use crate::codegen_cprover_gotoc::codegen::ty_stable::pointee_type_stable;
 use crate::codegen_cprover_gotoc::{GotocCtx, utils};
 use crate::intrinsics::Intrinsic;
 use crate::unwrap_or_return_codegen_unimplemented_stmt;
-use cbmc::MachineModel;
 use cbmc::goto_program::{
     ArithmeticOverflowResult, BinaryOperator, BuiltinFn, Expr, Location, Stmt, Type,
 };
@@ -16,7 +15,7 @@ use rustc_middle::ty::layout::ValidityRequirement;
 use rustc_smir::rustc_internal;
 use stable_mir::mir::mono::Instance;
 use stable_mir::mir::{BasicBlockIdx, Operand, Place};
-use stable_mir::ty::{FloatTy, GenericArgs, IntTy, RigidTy, Span, Ty, TyKind, UintTy};
+use stable_mir::ty::{GenericArgs, RigidTy, Span, Ty, TyKind, UintTy};
 use tracing::debug;
 
 pub struct SizeAlign {
@@ -1959,14 +1958,10 @@ impl GotocCtx<'_> {
             panic!("Expected intrinsic `{}` type to be `Float`, but found: `{:?}`", intrinsic, ty);
         };
         let mm = self.symbol_table.machine_model();
-        let (lower, upper) = match integral_ty {
-            RigidTy::Uint(uint_ty) => get_bounds_uint_expr(float_type, uint_ty, mm),
-            RigidTy::Int(int_ty) => get_bounds_int_expr(float_type, int_ty, mm),
-            _ => unreachable!(),
-        };
+        let in_range = utils::codegen_in_range_expr(&expr, float_type, integral_ty, mm);
 
         let range_check = self.codegen_assert_assume(
-            expr.clone().gt(lower).and(expr.clone().lt(upper)),
+            in_range,
             PropertyClass::ArithmeticOverflow,
             format!("{intrinsic}: attempt to convert a value out of range of the target integer")
                 .as_str(),
@@ -1992,280 +1987,4 @@ fn instance_args(instance: &Instance) -> GenericArgs {
         )
     };
     args
-}
-
-fn get_bounds_uint_expr(float_ty: FloatTy, uint_ty: UintTy, mm: &MachineModel) -> (Expr, Expr) {
-    match float_ty {
-        FloatTy::F32 => {
-            let (lower, upper) = get_bounds_f32_uint(uint_ty, mm);
-            (Expr::float_constant(lower), Expr::float_constant(upper))
-        }
-        FloatTy::F64 => {
-            let (lower, upper) = get_bounds_f64_uint(uint_ty, mm);
-            (Expr::double_constant(lower), Expr::double_constant(upper))
-        }
-        _ => unimplemented!(),
-    }
-}
-
-fn get_bounds_int_expr(float_ty: FloatTy, int_ty: IntTy, mm: &MachineModel) -> (Expr, Expr) {
-    match float_ty {
-        FloatTy::F32 => {
-            let (lower, upper) = get_bounds_f32_int(int_ty, mm);
-            (Expr::float_constant(lower), Expr::float_constant(upper))
-        }
-        FloatTy::F64 => {
-            let (lower, upper) = get_bounds_f64_int(int_ty, mm);
-            (Expr::double_constant(lower), Expr::double_constant(upper))
-        }
-        _ => unimplemented!(),
-    }
-}
-
-const F32_U_LOWER: [u8; 4] = [0x00, 0x00, 0x80, 0xBF]; // -1.0
-const F32_U8_UPPER: [u8; 4] = [0x00, 0x00, 0x80, 0x43]; // 256.0
-const F32_U16_UPPER: [u8; 4] = [0x00, 0x00, 0x80, 0x47]; // 65536.0
-const F32_U32_UPPER: [u8; 4] = [0x00, 0x00, 0x80, 0x4F]; // 4294967296.0
-const F32_U64_UPPER: [u8; 4] = [0x00, 0x00, 0x80, 0x5F]; // 18446744073709551616.0
-// The largest f32 value fits in a u128, so there is no upper bound
-const F32_U128_UPPER: [u8; 4] = [0x00, 0x00, 0x80, 0x7F]; // inf
-
-const F32_I8_LOWER: [u8; 4] = [0x00, 0x00, 0x01, 0xC3]; // -129.0
-const F32_I16_LOWER: [u8; 4] = [0x00, 0x01, 0x00, 0xC7]; // -32769.0
-const F32_I32_LOWER: [u8; 4] = [0x01, 0x00, 0x00, 0xCF]; // -2147483904.0
-const F32_I64_LOWER: [u8; 4] = [0x01, 0x00, 0x00, 0xDF]; // -9223373136366403584.0
-const F32_I128_LOWER: [u8; 4] = [0x01, 0x00, 0x00, 0xFF]; // -170141203742878835383357727663135391744.0
-const F32_I8_UPPER: [u8; 4] = [0x00, 0x00, 0x00, 0x43]; // 128.0
-const F32_I16_UPPER: [u8; 4] = [0x00, 0x00, 0x00, 0x47]; // 32768.0
-const F32_I32_UPPER: [u8; 4] = [0x00, 0x00, 0x00, 0x4F]; // 2147483648.0
-const F32_I64_UPPER: [u8; 4] = [0x00, 0x00, 0x00, 0x5F]; // 9223372036854775808.0
-const F32_I128_UPPER: [u8; 4] = [0x00, 0x00, 0x00, 0x7F]; // 170141183460469231731687303715884105728.0
-
-const F64_U_LOWER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0xBF]; // -1.0
-const F64_U8_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x70, 0x40]; // 256.0
-const F64_U16_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x40]; // 65536.0
-const F64_U32_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x41]; // 4294967296.0
-const F64_U64_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x43]; // 18446744073709551616.0
-const F64_U128_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x47]; // 340282366920938463463374607431768211456.0
-
-const F64_I8_LOWER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x60, 0xC0]; // -129.0
-const F64_I16_LOWER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x20, 0x00, 0xE0, 0xC0]; // -32769.0
-const F64_I32_LOWER: [u8; 8] = [0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0xE0, 0xC1]; // -2147483649.0
-const F64_I64_LOWER: [u8; 8] = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0xE0, 0xC3]; // -9223372036854777856.0
-const F64_I128_LOWER: [u8; 8] = [0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0xE0, 0xC7]; // -170141183460469269510619166673045815296.0
-const F64_I8_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0x40]; // 128.0
-const F64_I16_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xE0, 0x40]; // 32768.0
-const F64_I32_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xE0, 0x41]; // 2147483648.0
-const F64_I64_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xE0, 0x43]; // 9223372036854775808.0
-const F64_I128_UPPER: [u8; 8] = [0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xE0, 0x47]; // 170141183460469231731687303715884105728.0
-
-/// upper is the smallest `f32` that after truncation is strictly larger than u<N>::MAX
-/// lower is the largest `f32` that after truncation is strictly smaller than u<N>::MIN
-///
-/// For example, for N = 8, upper is 256.0 because the previous f32 (i.e.
-/// `256_f32.next_down()` which is 255.9999847412109375) when truncated is 255.0,
-/// which is not strictly larger than `u8::MAX`
-///
-/// For all bit-widths, lower is -1.0 because the next higher number, when
-/// truncated is -0.0 (or 0.0) which is not strictly smaller than `u<N>::MIN`
-fn get_bounds_f32_uint(uint_ty: UintTy, mm: &MachineModel) -> (f32, f32) {
-    let lower: f32 = f32::from_le_bytes(F32_U_LOWER);
-    let upper: f32 = match uint_ty {
-        UintTy::U8 => f32::from_le_bytes(F32_U8_UPPER),
-        UintTy::U16 => f32::from_le_bytes(F32_U16_UPPER),
-        UintTy::U32 => f32::from_le_bytes(F32_U32_UPPER),
-        UintTy::U64 => f32::from_le_bytes(F32_U64_UPPER),
-        UintTy::U128 => f32::from_le_bytes(F32_U128_UPPER),
-        UintTy::Usize => match mm.pointer_width {
-            32 => f32::from_le_bytes(F32_U32_UPPER),
-            64 => f32::from_le_bytes(F32_U64_UPPER),
-            _ => unreachable!(),
-        },
-    };
-    (lower, upper)
-}
-
-fn get_bounds_f64_uint(uint_ty: UintTy, mm: &MachineModel) -> (f64, f64) {
-    let lower = f64::from_le_bytes(F64_U_LOWER);
-    let upper = match uint_ty {
-        UintTy::U8 => f64::from_le_bytes(F64_U8_UPPER),
-        UintTy::U16 => f64::from_le_bytes(F64_U16_UPPER),
-        UintTy::U32 => f64::from_le_bytes(F64_U32_UPPER),
-        UintTy::U64 => f64::from_le_bytes(F64_U64_UPPER),
-        UintTy::U128 => f64::from_le_bytes(F64_U128_UPPER),
-        UintTy::Usize => match mm.pointer_width {
-            32 => f64::from_le_bytes(F64_U32_UPPER),
-            64 => f64::from_le_bytes(F64_U64_UPPER),
-            _ => unreachable!(),
-        },
-    };
-    (lower, upper)
-}
-
-/// upper is the smallest `f32` that after truncation is strictly larger than i<N>::MAX
-/// lower is the largest `f32` that after truncation is strictly smaller than i<N>::MIN
-///
-/// For example, for N = 16, upper is 32768.0 because the previous f32 (i.e.
-/// `32768_f32.next_down()`) when truncated is 32767,
-/// which is not strictly larger than `i16::MAX`
-///
-/// Note that all upper bound values are 2^(w-1) which can be precisely
-/// represented in f32 (verified using
-/// https://www.h-schmidt.net/FloatConverter/IEEE754.html)
-/// However, for lower bound values, which should be -2^(w-1)-1 (i.e.
-/// i<N>::MIN-1), not all of them can be represented in f32.
-/// For instance, for w = 32, -2^(31)-1 = -2,147,483,649, but this number does
-/// **not** have an f32 representation, and the next **smaller** number is
-/// -2,147,483,904. Note that CBMC for example uses the formula above which
-/// leads to bugs, e.g.: https://github.com/diffblue/cbmc/issues/8488
-fn get_bounds_f32_int(int_ty: IntTy, mm: &MachineModel) -> (f32, f32) {
-    let lower = match int_ty {
-        IntTy::I8 => f32::from_le_bytes(F32_I8_LOWER),
-        IntTy::I16 => f32::from_le_bytes(F32_I16_LOWER),
-        IntTy::I32 => f32::from_le_bytes(F32_I32_LOWER),
-        IntTy::I64 => f32::from_le_bytes(F32_I64_LOWER),
-        IntTy::I128 => f32::from_le_bytes(F32_I128_LOWER),
-        IntTy::Isize => match mm.pointer_width {
-            32 => f32::from_le_bytes(F32_I32_LOWER),
-            64 => f32::from_le_bytes(F32_I64_LOWER),
-            _ => unreachable!(),
-        },
-    };
-
-    let upper = match int_ty {
-        IntTy::I8 => f32::from_le_bytes(F32_I8_UPPER),
-        IntTy::I16 => f32::from_le_bytes(F32_I16_UPPER),
-        IntTy::I32 => f32::from_le_bytes(F32_I32_UPPER),
-        IntTy::I64 => f32::from_le_bytes(F32_I64_UPPER),
-        IntTy::I128 => f32::from_le_bytes(F32_I128_UPPER),
-        IntTy::Isize => match mm.pointer_width {
-            32 => f32::from_le_bytes(F32_I32_UPPER),
-            64 => f32::from_le_bytes(F32_I64_UPPER),
-            _ => unreachable!(),
-        },
-    };
-    (lower, upper)
-}
-
-fn get_bounds_f64_int(int_ty: IntTy, mm: &MachineModel) -> (f64, f64) {
-    let lower = match int_ty {
-        IntTy::I8 => f64::from_le_bytes(F64_I8_LOWER),
-        IntTy::I16 => f64::from_le_bytes(F64_I16_LOWER),
-        IntTy::I32 => f64::from_le_bytes(F64_I32_LOWER),
-        IntTy::I64 => f64::from_le_bytes(F64_I64_LOWER),
-        IntTy::I128 => f64::from_le_bytes(F64_I128_LOWER),
-        IntTy::Isize => match mm.pointer_width {
-            32 => f64::from_le_bytes(F64_I32_LOWER),
-            64 => f64::from_le_bytes(F64_I64_LOWER),
-            _ => unreachable!(),
-        },
-    };
-    let upper = match int_ty {
-        IntTy::I8 => f64::from_le_bytes(F64_I8_UPPER),
-        IntTy::I16 => f64::from_le_bytes(F64_I16_UPPER),
-        IntTy::I32 => f64::from_le_bytes(F64_I32_UPPER),
-        IntTy::I64 => f64::from_le_bytes(F64_I64_UPPER),
-        IntTy::I128 => f64::from_le_bytes(F64_I128_UPPER),
-        IntTy::Isize => match mm.pointer_width {
-            32 => f64::from_le_bytes(F64_I32_UPPER),
-            64 => f64::from_le_bytes(F64_I64_UPPER),
-            _ => unreachable!(),
-        },
-    };
-    (lower, upper)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use num::BigInt;
-    use num::FromPrimitive;
-
-    macro_rules! check_lower_f32 {
-        ($val:ident, $min:expr) => {
-            let f = f32::from_le_bytes($val);
-            assert!(BigInt::from_f32(f.trunc()).unwrap() < BigInt::from($min));
-            assert!(BigInt::from_f32(f.next_up().trunc()).unwrap() >= BigInt::from($min));
-        };
-    }
-
-    macro_rules! check_upper_f32 {
-        ($val:ident, $max:expr) => {
-            let f = f32::from_le_bytes($val);
-            assert!(BigInt::from_f32(f.trunc()).unwrap() > BigInt::from($max));
-            assert!(BigInt::from_f32(f.next_down().trunc()).unwrap() <= BigInt::from($max));
-        };
-    }
-
-    #[test]
-    fn check_f32_bounds() {
-        // check that the bounds are correct, i.e. that for lower (upper) bounds:
-        //   1. the value when truncated is strictly smaller (larger) than {i, u}<N>::MIN ({i, u}<N>::MAX)
-        //   2. the next higher (lower) value when truncated is greater (smaller) than or equal to {i, u}<N>::MIN ({i, u}<N>::MAX)
-
-        check_lower_f32!(F32_U_LOWER, u8::MIN);
-
-        check_upper_f32!(F32_U8_UPPER, u8::MAX);
-        check_upper_f32!(F32_U16_UPPER, u16::MAX);
-        check_upper_f32!(F32_U32_UPPER, u32::MAX);
-        check_upper_f32!(F32_U64_UPPER, u64::MAX);
-        // 128 is not needed because the upper bounds is infinity
-        // Instead, check that `u128::MAX` is larger than the largest f32 value
-        assert!(f32::MAX < u128::MAX as f32);
-
-        check_lower_f32!(F32_I8_LOWER, i8::MIN);
-        check_lower_f32!(F32_I16_LOWER, i16::MIN);
-        check_lower_f32!(F32_I32_LOWER, i32::MIN);
-        check_lower_f32!(F32_I64_LOWER, i64::MIN);
-        check_lower_f32!(F32_I128_LOWER, i128::MIN);
-
-        check_upper_f32!(F32_I8_UPPER, i8::MAX);
-        check_upper_f32!(F32_I16_UPPER, i16::MAX);
-        check_upper_f32!(F32_I32_UPPER, i32::MAX);
-        check_upper_f32!(F32_I64_UPPER, i64::MAX);
-        check_upper_f32!(F32_I128_UPPER, i128::MAX);
-    }
-
-    macro_rules! check_lower_f64 {
-        ($val:ident, $min:expr) => {
-            let f = f64::from_le_bytes($val);
-            assert!(BigInt::from_f64(f.trunc()).unwrap() < BigInt::from($min));
-            assert!(BigInt::from_f64(f.next_up().trunc()).unwrap() >= BigInt::from($min));
-        };
-    }
-
-    macro_rules! check_upper_f64 {
-        ($val:ident, $max:expr) => {
-            let f = f64::from_le_bytes($val);
-            assert!(BigInt::from_f64(f.trunc()).unwrap() > BigInt::from($max));
-            assert!(BigInt::from_f64(f.next_down().trunc()).unwrap() <= BigInt::from($max));
-        };
-    }
-
-    #[test]
-    fn check_f64_bounds() {
-        // check that the bounds are correct, i.e. that for lower (upper) bounds:
-        //   1. the value when truncated is strictly smaller (larger) than {i, u}<N>::MIN ({i, u}<N>::MAX)
-        //   2. the next higher (lower) value when truncated is greater (smaller) than or equal to {i, u}<N>::MIN ({i, u}<N>::MAX)
-
-        check_lower_f64!(F64_U_LOWER, u8::MIN);
-
-        check_upper_f64!(F64_U8_UPPER, u8::MAX);
-        check_upper_f64!(F64_U16_UPPER, u16::MAX);
-        check_upper_f64!(F64_U32_UPPER, u32::MAX);
-        check_upper_f64!(F64_U64_UPPER, u64::MAX);
-        check_upper_f64!(F64_U128_UPPER, u128::MAX);
-
-        check_lower_f64!(F64_I8_LOWER, i8::MIN);
-        check_lower_f64!(F64_I16_LOWER, i16::MIN);
-        check_lower_f64!(F64_I32_LOWER, i32::MIN);
-        check_lower_f64!(F64_I64_LOWER, i64::MIN);
-        check_lower_f64!(F64_I128_LOWER, i128::MIN);
-
-        check_upper_f64!(F64_I8_UPPER, i8::MAX);
-        check_upper_f64!(F64_I16_UPPER, i16::MAX);
-        check_upper_f64!(F64_I32_UPPER, i32::MAX);
-        check_upper_f64!(F64_I64_UPPER, i64::MAX);
-        check_upper_f64!(F64_I128_UPPER, i128::MAX);
-    }
 }
