+++ kani-compiler/src/codegen_cprover_gotoc/utils/mod.rs
@@ -4,12 +4,14 @@
 //! This module provides utils used across Kani
 
 mod debug;
+mod float_utils;
 mod names;
 #[allow(clippy::module_inception)]
 mod utils;
 
 // TODO clean this up
 
+pub use float_utils::*;
 pub use names::*;
 pub use utils::*;
